// src/enums/category.ts
export enum CategoryEnum {
  Brasil = 12,
  Cultura = 10,
  Economia = 1,
  Educacao = 8,
  Esportes = 2,
  MidiCia = 35,
  Mundo = 15,
  Rio = 13,
  Saude = 3,
  Turismo = 6,
}

export enum CategorySlugEnum {
  Brasil = 'brasil',
  Cultura = 'cultura',
  Economia = 'economia',
  Educacao = 'educacao',
  Esportes = 'esportes',
  MidiCia = 'midia-cia',
  Mundo = 'mundo',
  Rio = 'rio',
  Saude = 'saude',
  Turismo = 'turismo',
}

export enum CategoryTitleEnum {
  Brasil = 'Brasil',
  Cultura = 'Cultura',
  Economia = 'Economia',
  Educacao = 'Educação',
  Esportes = 'Esportes',
  MidiCia = 'Mídia & Cia',
  Mundo = 'Mundo',
  Rio = 'Rio',
  Saude = 'Saúde',
  Turismo = 'Turismo',
}